Migrate rsfmt into sfmt v0.2
