+++ check/SFMT_19937.txt
