+++ check/init.c
