--- rsfmt/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "rsfmt"
-version = "0.1.0"
-authors = ["Toshiki Teramura <toshiki.teramura@gmail.com>"]
-
-[dependencies]
-rand = "0.4"
