--- sfmt-sys/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "sfmt-sys"
-version = "0.1.0"
-authors = ["Toshiki Teramura <toshiki.teramura@gmail.com>"]
-description = "Rust wrapper of SFMT"
-documentation = "https://docs.rs/sfmt-sys/"
-repository = "https://github.com/termoshtt/rust-sfmt"
-keywords = ["random", "rng"]
-categories = ["algorithms"]
-license = "MIT"
-links = "sfmt"
